mpd-monitor/requirements.txt
python-mpd2==1.0.0
uPnPClient==0.0.8
